+++ api/src/functions/auth.ts
@@ -33,9 +33,9 @@ export const handler = async (
       // for security reasons you may want to be vague here rather than expose
       // the fact that the email address wasn't found (prevents fishing for
       // valid email addresses)
-      usernameNotFound: 'User Not Found',
+      usernameNotFound: 'Not Found',
       // if the user somehow gets around client validation
-      usernameRequired: 'Information Required',
+      usernameRequired: 'Required',
     },
   }
 
@@ -56,12 +56,12 @@ export const handler = async (
     },
 
     errors: {
-      usernameOrPasswordMissing: 'Both username and password are required',
-      usernameNotFound: 'User Not Found',
+      usernameOrPasswordMissing: 'All Info Required',
+      usernameNotFound: 'Not Found',
       // For security reasons you may want to make this the same as the
       // usernameNotFound error so that a malicious user can't use the error
       // to narrow down if it's the username or password that's incorrect
-      incorrectPassword: 'Incorrect Credentials',
+      incorrectPassword: 'Incorrect',
     },
 
     // How long a user will remain logged in, in seconds